+++ database/migrations/2025_02_05_093701_create_final_semester_enrollments_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(FinalSessionEnrollment::class)->constrained();
             $table->foreignIdFor(Semester::class)->constrained();
-            $table->unsignedTinyInteger('result_count')->default(0);
-            $table->unsignedTinyInteger('credit_unit_sum')->default(0);
+            $table->unsignedSmallInteger('result_count')->default(0);
+            $table->unsignedSmallInteger('credit_unit_sum')->default(0);
             $table->unsignedSmallInteger('grade_point_sum')->default(0);
             $table->unsignedSmallInteger('grade_point_average')->default(0);
             $table->timestamps();
